상품추가(이미지X) , 상품수정(수정후 이동 경로 변경 필요)

+++ src/main/java/com/example/kuly2/admin/Dto/ProductDto.java
@@ -17,7 +17,7 @@ import java.sql.Timestamp;
 @NoArgsConstructor
 public class ProductDto {
     @Id
-    private int product_no;
+    private Long product_no;
     private String product_name;
     private int product_price;
     private String product_information;
@@ -29,7 +29,7 @@ public class ProductDto {
 
 
     @Builder
-    public ProductDto(int product_no, String product_name, int product_price,
+    public ProductDto(Long product_no, String product_name, int product_price,
                          String product_information, String product_photo,
                          String product_description, int product_rate ,String product_adder,
                          Timestamp product_regdate) {
